Added during_move and after_move settings
Being able to call a function during a move and after a move was needed for me, so I wanted to contribute an update.  This could be used for instance when sorting rows and showing a row number.  When moving the rows during sort, the row numbers could temporarily hide, and when sorting is finished, reappear with the new sorting.